@@ -265,7 +265,8 @@ abstract class Writer implements IWriter
     {
 
         //Output the survey.
-        $headers = array();
+        $headers = [];
+        $fieldNames = [];
         if ($bOutputHeaders) {
             foreach ($oOptions->selectedColumns as $sColumn) {
                 //Survey question field, $column value is a field name from the getFieldMap function.
@@ -285,6 +286,7 @@ abstract class Writer implements IWriter
                         break;
                 }
                 $headers[] = $value;
+                $fieldNames[] = $sColumn;
             }
         }
         //Output the results.
@@ -314,7 +316,7 @@ abstract class Writer implements IWriter
             $oResponse->decrypt();
             $aResponse = array_merge($aResponse, $oResponse->attributes);
 
-            $elementArray = array();
+            $elementArray = [];
 
             foreach ($oOptions->selectedColumns as $column) {
                 $value = $aResponse[$column];
@@ -334,9 +336,9 @@ abstract class Writer implements IWriter
                 }
             }
             if ($oOptions->output == 'display') {
-                $this->outputRecord($headers, $elementArray, $oOptions);
+                $this->outputRecord($headers, $elementArray, $oOptions, $fieldNames);
             } else {
-                $sFile .= $this->outputRecord($headers, $elementArray, $oOptions);
+                $sFile .= $this->outputRecord($headers, $elementArray, $oOptions, $fieldNames);
             }
         }
         return $sFile;
@@ -360,6 +362,7 @@ abstract class Writer implements IWriter
      * @param array $headers
      * @param array $values
      * @param FormattingOptions $oOptions
+     * @param array $fieldNames
      */
-    abstract protected function outputRecord($headers, $values, FormattingOptions $oOptions);
+    abstract protected function outputRecord($headers, $values, FormattingOptions $oOptions, $fieldNames = []);
 }
