@@ -9,9 +9,13 @@ $rest = [
                 'name' => 'Survey',
                 'description' => 'Survey',
             ],
+            'auth' => [
+                'name' => 'Auth',
+                'description' => 'Auth',
+            ],
             'session' => [
                 'name' => 'Session',
-                'description' => 'Session',
+                'description' => 'Session (deprecated - use auth)',
             ],
             'survey-group' => [
                 'name' => 'Survey Group',
@@ -20,7 +24,11 @@ $rest = [
             'site-settings' => [
                 'name' => 'Site Settings',
                 'description' => 'Site Settings',
-            ]
+            ],
+            'user' => [
+                'name' => 'User',
+                'description' => 'User',
+            ],
         ]
     ]
 ];
@@ -28,7 +36,7 @@ $rest = [
 return array_merge(
     $rest,
     include_once __DIR__ . '/v1/survey.php',
-    include_once __DIR__ . '/v1/session.php',
+    include_once __DIR__ . '/v1/auth.php',
     include_once __DIR__ . '/v1/survey-group.php',
     include_once __DIR__ . '/v1/user.php',
     include_once __DIR__ . '/v1/site-settings.php',
