@@ -1,8 +1,10 @@
 <?php
 
 use LimeSurvey\Api\Command\V1\{
-    SessionKeyCreate,
-    SessionKeyRelease
+    AuthSessionCreate,
+    AuthTokenSimpleCreate,
+    AuthTokenSimpleRefresh,
+    AuthTokenSimpleRelease
 };
 use LimeSurvey\Api\Rest\V1\SchemaFactory\{
     SchemaFactoryError,
@@ -13,10 +15,10 @@ $errorSchema = (new SchemaFactoryError)->make();
 
 $rest = [];
 
-$rest['v1/session'] = [
+$rest['v1/auth'] = [
     'POST' => [
         'description' => 'Generate new authentication token',
-        'commandClass' => SessionKeyCreate::class,
+        'commandClass' => AuthTokenSimpleCreate::class,
         'params' => [
             'username' => ['src' => 'form'],
             'password' => ['src' => 'form']
@@ -36,10 +38,30 @@ $rest['v1/session'] = [
             ]
         ]
     ],
+    'PUT' => [
+        'description' => 'Refresh authentication token',
+        'commandClass' => AuthTokenSimpleRefresh::class,
+        'auth' => true,
+        'params' => [],
+        'bodyParams' => [],
+        'responses' => [
+            'success' => [
+                'code' => 200,
+                'description' => 'Success - returns string access token for use in header '
+                . '"Authorization: Bearer $token"',
+                'schema' => (new SchemaFactoryAuthToken)->make()
+            ],
+            'unauthorized' => [
+                'code' => 401,
+                'description' => 'Unauthorized',
+                'schema' => $errorSchema
+            ]
+        ]
+    ],
     'DELETE' => [
         'description' => 'Destroy currently used authentication token',
-        'commandClass' => SessionKeyRelease::class,
-        'auth' => 'session',
+        'commandClass' => AuthTokenSimpleRelease::class,
+        'auth' => true,
         'params' => [],
         'bodyParams' => [],
         'responses' => [
@@ -55,5 +77,11 @@ $rest['v1/session'] = [
         ]
     ]
 ];
+
+// Add session endpoints to auth endpoints for backward compatibility
+// - can remove this once the survey template functionality is calling
+// - /auth instead of /session
+$rest['v1/session'] = $rest['v1/auth'];
+$rest['v1/session']['POST']['commandClass'] = AuthSessionCreate::class;
 
 return $rest;
