@@ -24,7 +24,7 @@ $rest['v1/survey'] = [
     'GET' => [
         'description' => 'Survey list',
         'commandClass' => SurveyList::class,
-        'auth' => 'session',
+        'auth' => true,
         'params' => [
             'pageSize' => ['type' => 'int'],
             'page' => ['type' => 'int']
@@ -50,7 +50,7 @@ $rest['v1/survey-detail/$id'] = [
         'tag' => 'survey',
         'description' => 'Survey detail',
         'commandClass' => SurveyDetail::class,
-        'auth' => 'session',
+        'auth' => true,
         'responses' => [
             'success' => [
                 'code' => 200,
@@ -74,7 +74,7 @@ $rest['v1/survey-detail/$id'] = [
         'tag' => 'survey',
         'description' => 'Survey update via RFC 6902 based patch',
         'commandClass' => SurveyPatch::class,
-        'auth' => 'session',
+        'auth' => true,
         'example' => __DIR__ . '/example/survey-patch-all.json',
         'schema' => $surveyPatchSchema,
         'responses' => [
@@ -104,7 +104,7 @@ $rest['v1/survey-template/$id'] = [
         'tag' => 'survey',
         'description' => 'Survey template',
         'commandClass' => SurveyTemplate::class,
-        'auth' => 'session',
+        'auth' => true,
         'responses' => [
             'success' => [
                 'code' => 200,
@@ -123,7 +123,7 @@ $rest['v1/survey-template/$id'] = [
         'tag' => 'survey',
         'description' => 'Survey template',
         'commandClass' => SurveyTemplate::class,
-        'auth' => 'session',
+        'auth' => true,
         'example' => __DIR__ . '/example/survey-post-template.json',
         'schema' => $surveyTemplateSchema,
         'responses' => [
