@@ -13,7 +13,8 @@
  *
  */
 
-use LimeSurvey\Api\Auth\AuthSession;
+use LimeSurvey\DI;
+use LimeSurvey\Api\Authentication\AuthenticationTokenSimple;
 
 class EditorLinkController extends LSYii_Controller
 {
@@ -50,7 +51,7 @@ class EditorLinkController extends LSYii_Controller
      */
     public function run($action)
     {
-        $this->setAuthInitCookie();
+        $this->setAuthenticationInitCookie();
         $editorUrl = Yii::app()->request->getQuery(
             'url',
             Yii::app()->request->baseUrl
@@ -66,27 +67,28 @@ class EditorLinkController extends LSYii_Controller
      *
      * @return void
      */
-    private function setAuthInitCookie()
+    private function setAuthenticationInitCookie()
     {
+        $diContainer = DI::getContainer();
+
         $cookieName = 'LS_AUTH_INIT';
 
-        $authSession = new AuthSession();
-        $session = $authSession->createSession(
-            Yii::app()->session['user']
+        $authTokenSimple = $diContainer->get(
+            AuthenticationTokenSimple::class
         );
-
-        $sessionExpires = new \DateTime(
-            date('c', $session->expire)
+        $session = $authTokenSimple->createSession(
+            Yii::app()->session['user']
         );
 
         /** @var \LSYii_Application */
         $app = \Yii::app();
 
-        $cookieDataJson = json_encode([
-            'token' => $session->id,
-            'userId' => $app->user->id,
-            'expires' => $sessionExpires->format('Y-m-d\TH:i:s.000\Z')
-        ]);
+        $cookieDataJson = json_encode(
+            $authTokenSimple->getTokenData(
+                $session,
+                $app->user->getId()
+            )
+        );
 
         $cookie = new CHttpCookie($cookieName, $cookieDataJson);
         $cookie->expire = time() + (60 * 2); // 2 minutes
