@@ -12,6 +12,16 @@ class PasswordRequirement extends \LimeSurvey\PluginManager\PluginBase
      */
     protected $storage = 'DbStorage';
 
+    const DEFAULT_NEEDS_NUMBER = true;
+    const DEFAULT_NEEDS_UPPERCASE = true;
+    const DEFAULT_NEEDS_NON_ALPHANUMERIC = false;
+    const DEFAULT_MINIMUM_SIZE = 12;
+    const DEFAULT_SURVEY_SAVE_ACTIVE = false;
+    const DEFAULT_SURVEY_SAVE_NEEDS_NUMBER = false;
+    const DEFAULT_SURVEY_SAVE_NEEDS_UPPERCASE = false;
+    const DEFAULT_SURVEY_SAVE_NEEDS_NON_ALPHANUMERIC = false;
+    const DEFAULT_SURVEY_SAVE_MINIMUM_SIZE = 8;
+
     /** @inheritdoc this plugin didn't have any public method */
     public $allowedPublicMethods = array();
 
@@ -27,22 +37,21 @@ class PasswordRequirement extends \LimeSurvey\PluginManager\PluginBase
         'needsNumber' => array(
             'label' => 'Require at least one digit',
             'type' => 'checkbox',
-            'default' => true,
+            'default' => self::DEFAULT_NEEDS_NUMBER,
         ),
         'needsUppercase' => array(
             'label' => 'Require at least one uppercase character',
             'type' => 'checkbox',
-            'default' => true,
+            'default' => self::DEFAULT_NEEDS_UPPERCASE,
         ),
         'needsNonAlphanumeric' => array(
             'label' => 'Require at least one special character',
             'type' => 'checkbox',
-            'default' => false,
+            'default' => self::DEFAULT_NEEDS_NON_ALPHANUMERIC,
         ),
         'minimumSize' => array(
             'label' => 'Minimum password length',
             'type' => 'int',
-            'default' => 12,
         ),
         'surveyPart' => array(
             'content' => 'Password requirements for “Save and return later” feature',
@@ -55,27 +64,26 @@ class PasswordRequirement extends \LimeSurvey\PluginManager\PluginBase
         'surveySaveActive' => array(
             'type' => 'boolean',
             'label' => 'Check password when saving survey',
-            'default' => false,
+            'default' => self::DEFAULT_SURVEY_SAVE_ACTIVE,
         ),
         'surveySaveNeedsNumber' => array(
             'label' => 'Require at least one digit',
             'type' => 'checkbox',
-            'default' => false,
+            'default' => self::DEFAULT_SURVEY_SAVE_NEEDS_NUMBER,
         ),
         'surveySaveNeedsUppercase' => array(
             'label' => 'Require at least one uppercase character',
             'type' => 'checkbox',
-            'default' => false,
+            'default' => self::DEFAULT_SURVEY_SAVE_NEEDS_UPPERCASE,
         ),
         'surveySaveNeedsNonAlphanumeric' => array(
             'label' => 'Require at least one special character',
             'type' => 'checkbox',
-            'default' => false,
+            'default' => self::DEFAULT_SURVEY_SAVE_NEEDS_NON_ALPHANUMERIC,
         ),
         'surveySaveMinimumSize' => array(
             'label' => 'Minimum password length',
             'type' => 'int',
-            'default' => 8,
         ),
     ];
     /**
@@ -95,10 +103,10 @@ class PasswordRequirement extends \LimeSurvey\PluginManager\PluginBase
         $password = $oEvent->get('password');
         $errors = $this->checkValidityOfPassword(
             $password,
-            $this->get('needsNumber', null, null, true),
-            $this->get('needsUppercase', null, null, true),
-            $this->get('needsNonAlphanumeric', null, null, false),
-            $this->get('minimumSize', null, null, 8)
+            $this->get('needsNumber', null, null, self::DEFAULT_NEEDS_NUMBER),
+            $this->get('needsUppercase', null, null, self::DEFAULT_NEEDS_UPPERCASE),
+            $this->get('needsNonAlphanumeric', null, null, self::DEFAULT_NEEDS_NON_ALPHANUMERIC),
+            $this->get('minimumSize', null, null, self::DEFAULT_MINIMUM_SIZE)
         );
         if ($errors) {
             $oEvent->set('passwordOk', false);
@@ -113,7 +121,7 @@ class PasswordRequirement extends \LimeSurvey\PluginManager\PluginBase
      * */
     public function validateSaveSurveyForm()
     {
-        if (!$this->get('surveySaveActive', null, null, false)) {
+        if (!$this->get('surveySaveActive', null, null, self::DEFAULT_SURVEY_SAVE_ACTIVE)) {
             return;
         }
         $event = $this->getEvent();
@@ -131,10 +139,10 @@ class PasswordRequirement extends \LimeSurvey\PluginManager\PluginBase
         $password = $saveData['clearpassword'];
         $errors = $this->checkValidityOfPassword(
             $password,
-            $this->get('surveySaveNeedsNumber', null, null, false),
-            $this->get('surveySaveNeedsUppercase', null, null, false),
-            $this->get('surveySaveNeedsNonAlphanumeric', null, null, false),
-            $this->get('surveySaveMinimumSize', null, null, 8)
+            $this->get('surveySaveNeedsNumber', null, null, self::DEFAULT_SURVEY_SAVE_NEEDS_NUMBER),
+            $this->get('surveySaveNeedsUppercase', null, null, self::DEFAULT_SURVEY_SAVE_NEEDS_UPPERCASE),
+            $this->get('surveySaveNeedsNonAlphanumeric', null, null, self::DEFAULT_SURVEY_SAVE_NEEDS_NON_ALPHANUMERIC),
+            $this->get('surveySaveMinimumSize', null, null, self::DEFAULT_SURVEY_SAVE_MINIMUM_SIZE)
         );
         if (empty($errors)) {
             return;
@@ -176,10 +184,11 @@ class PasswordRequirement extends \LimeSurvey\PluginManager\PluginBase
         $oEvent = $this->getEvent();
         $targetSize = $oEvent->get('targetSize', 8);
 
-        $targetSize = $targetSize < $this->get('minimumSize', null, null, 12) ? $this->get('minimumSize', null, null, 12) : $targetSize;
-        $uppercase = $this->get('needsUppercase', null, null, true);
-        $numeric = $this->get('needsNumber', null, null, true);
-        $nonAlpha = $this->get('needsNonAlphanumeric', null, null, false);
+        $minimumSize = $this->get('minimumSize', null, null, self::DEFAULT_MINIMUM_SIZE);
+        $targetSize = $targetSize < $minimumSize ? $minimumSize : $targetSize;
+        $uppercase = $this->get('needsUppercase', null, null, self::DEFAULT_NEEDS_UPPERCASE);
+        $numeric = $this->get('needsNumber', null, null, self::DEFAULT_NEEDS_NUMBER);
+        $nonAlpha = $this->get('needsNonAlphanumeric', null, null, self::DEFAULT_NEEDS_NON_ALPHANUMERIC);
 
         $randomPassword = $this->getRandomString($targetSize, $uppercase, $numeric, $nonAlpha);
         
@@ -199,12 +208,14 @@ class PasswordRequirement extends \LimeSurvey\PluginManager\PluginBase
         $settings['needsUppercase']['label'] = $this->gT("Require at least one uppercase character");
         $settings['needsNonAlphanumeric']['label'] = $this->gT("Require at least one special character");
         $settings['minimumSize']['label'] = $this->gT("Minimum password length");
+        $settings['minimumSize']['help'] = sprintf(gT('Default value will be %d if left blank'), self::DEFAULT_MINIMUM_SIZE);
         $settings['surveyPart']['content'] = $this->gT("Password requirements for “Save and return later” feature");
         $settings['surveySaveActive']['label'] = $this->gT("Check password when use “Save and return later” feature");
         $settings['surveySaveNeedsNumber']['label'] = $this->gT("Require at least one digit");
         $settings['surveySaveNeedsUppercase']['label'] = $this->gT("Require at least one uppercase character");
         $settings['surveySaveNeedsNonAlphanumeric']['label'] = $this->gT("Require at least one special character");
         $settings['surveySaveMinimumSize']['label'] = $this->gT("Minimum password length");
+        $settings['surveySaveMinimumSize']['help'] = sprintf(gT('Default value will be %d if left blank'), self::DEFAULT_SURVEY_SAVE_MINIMUM_SIZE);
         return $settings;
     }
 
