@@ -2,7 +2,7 @@
 
 namespace LimeSurvey\Api\Command\V1;
 
-use LimeSurvey\Api\Auth\AuthSession;
+use LimeSurvey\Api\Authentication\AuthenticationTokenSimple;
 use LimeSurvey\Api\Command\{
     CommandInterface,
     Request\Request,
@@ -10,22 +10,18 @@ use LimeSurvey\Api\Command\{
     Response\ResponseFactory
 };
 
-class SessionKeyRelease implements CommandInterface
+class AuthTokenSimpleRelease implements CommandInterface
 {
-    protected AuthSession $authSession;
     protected ResponseFactory $responseFactory;
 
     /**
      * Constructor
      *
-     * @param AuthSession $authSession
      * @param ResponseFactory $responseFactory
      */
     public function __construct(
-        AuthSession $authSession,
         ResponseFactory $responseFactory
     ) {
-        $this->authSession = $authSession;
         $this->responseFactory = $responseFactory;
     }
 
@@ -37,10 +33,6 @@ class SessionKeyRelease implements CommandInterface
      */
     public function run(Request $request)
     {
-        $this->authSession->doLogout(
-            $request
-            ->getData('sessionKey')
-        );
         return $this->responseFactory
             ->makeSuccess('OK');
     }
