@@ -10,7 +10,6 @@ use LimeSurvey\Api\Command\{
     Response\Response,
     Response\ResponseFactory
 };
-use LimeSurvey\Api\Auth\AuthSession;
 use LimeSurvey\Api\Command\Mixin\Auth\AuthPermissionTrait;
 
 class SurveyGroupList implements CommandInterface
@@ -18,7 +17,6 @@ class SurveyGroupList implements CommandInterface
     use AuthPermissionTrait;
 
     protected SurveysGroups $surveyGroup;
-    protected AuthSession $authSession;
     protected TransformerOutputSurveyGroup $transformerOutputSurveyGroup;
     protected ResponseFactory $responseFactory;
 
@@ -26,18 +24,16 @@ class SurveyGroupList implements CommandInterface
      * Constructor
      *
      * @param SurveysGroups $surveyGroup
-     * @param AuthSession $authSession
+     * @param AuthTokenSimple $auth
      * @param TransformerOutputSurveyGroup $transformerOutputSurveyGroup
      * @param ResponseFactory $responseFactory
      */
     public function __construct(
         SurveysGroups $surveyGroup,
-        AuthSession $authSession,
         TransformerOutputSurveyGroup $transformerOutputSurveyGroup,
         ResponseFactory $responseFactory
     ) {
         $this->surveyGroup = $surveyGroup;
-        $this->authSession = $authSession;
         $this->transformerOutputSurveyGroup = $transformerOutputSurveyGroup;
         $this->responseFactory = $responseFactory;
     }
@@ -50,16 +46,6 @@ class SurveyGroupList implements CommandInterface
      */
     public function run(Request $request)
     {
-        $sessionKey = (string) $request->getData('sessionKey');
-
-        if (
-            !$this->authSession
-                ->checkKey($sessionKey)
-        ) {
-            return $this->responseFactory
-                ->makeErrorUnauthorised();
-        }
-
         $dataProvider = $this->surveyGroup->search();
 
         $data = $this->transformerOutputSurveyGroup
