@@ -11,7 +11,6 @@ use LimeSurvey\Api\Command\{
     Response\Response,
     Response\ResponseFactory
 };
-use LimeSurvey\Api\Auth\AuthSession;
 use LimeSurvey\Api\Command\Mixin\Auth\AuthPermissionTrait;
 
 class UserList implements CommandInterface
@@ -20,7 +19,6 @@ class UserList implements CommandInterface
 
     protected User $user;
     protected Permission $permission;
-    protected AuthSession $authSession;
     protected TransformerOutputSurveyOwner $transformerOutputSurveyOwner;
     protected ResponseFactory $responseFactory;
 
@@ -29,20 +27,18 @@ class UserList implements CommandInterface
      *
      * @param User $user
      * @param Permission $permission
-     * @param AuthSession $authSession
+     * @param AuthTokenSimple $auth
      * @param TransformerOutputSurveyOwner $transformerOutputSurveyOwner
      * @param ResponseFactory $responseFactory
      */
     public function __construct(
         User $user,
         Permission $permission,
-        AuthSession $authSession,
         TransformerOutputSurveyOwner $transformerOutputSurveyOwner,
         ResponseFactory $responseFactory
     ) {
         $this->user = $user;
         $this->permission = $permission;
-        $this->authSession = $authSession;
         $this->transformerOutputSurveyOwner = $transformerOutputSurveyOwner;
         $this->responseFactory = $responseFactory;
     }
@@ -55,16 +51,6 @@ class UserList implements CommandInterface
      */
     public function run(Request $request)
     {
-        $sessionKey = (string) $request->getData('sessionKey');
-
-        if (
-            !$this->authSession
-                ->checkKey($sessionKey)
-        ) {
-            return $this->responseFactory
-                ->makeErrorUnauthorised();
-        }
-
         if (!$this->permission->hasGlobalPermission('users', 'read')) {
             return $this->responseFactory
                 ->makeErrorForbidden();
