@@ -15,9 +15,9 @@
         <tr class="active">
             <th colspan='4' align='center' style='text-align: center; '>
                 <strong>
-                    <?php echo sprintf(gT("Summary for %s"),$outputs['qtitle']); ?>
+                    <?php echo sprintf(gT("Summary for %s"), $outputs['qtitle']); ?>
                 </strong>
-                <button class="float-end action_js_export_to_pdf btn btn-outline-secondary btn-sm d-print-none" data-question-id="quid_<?php echo $outputs['parentqid'];?>" data-bs-toggle="tooltip" title="<?php eT('Export this question to PDF.'); ?>">
+                <button class="float-end action_js_export_to_pdf btn btn-outline-secondary btn-sm d-print-none" data-question-id="quid_<?php echo $outputs['parentqid'];?>" data-bs-toggle="tooltip" title="<?php eT('Export this question to PDF.'); ?>" onclick="return false;">
                     <i class="ri-file-pdf-line"></i>
                 </button>
             </th>
@@ -36,19 +36,19 @@
                 </strong>
             </th>
 
-            <?php if ($bShowCount  = true): ?>
+            <?php if ($bShowCount  = true) : ?>
                 <th width='' align='center' >
                     <strong><?php eT("Count"); ?></strong>
                 </th>
             <?php endif;?>
 
-            <?php if ($bShowPercentage  = true): ?>
-                <th width='' align='center' <?=(!$bSum ? 'colspan="2"' :'')?>>
+            <?php if ($bShowPercentage  = true) : ?>
+                <th width='' align='center' <?=(!$bSum ? 'colspan="2"' : '')?>>
                     <strong><?php eT("Gross percentage");?></strong>
                 </th>
             <?php endif;?>
 
-            <?php if($bSum): ?>
+            <?php if ($bSum) : ?>
                 <th width='' align='center' >
                     <strong>
                         <?php eT("Top 2, Middle, Bottom 2");?>
