@@ -1,10 +1,12 @@
 <?php
 // new question editor btn
-$this->renderPartial('/surveyAdministration/partial/topbar/_newQuestionEditorBtn', [
-    'editorUrl' => $editorUrl,
-    'enableEditorButton' => $enableEditorButton,
-    'editorEnabled' => $editorEnabled
-]);
+if (isset($editorUrl)) {
+    $this->renderPartial('/surveyAdministration/partial/topbar/_newQuestionEditorBtn', [
+        'editorUrl' => $editorUrl,
+        'enableEditorButton' => $enableEditorButton,
+        'editorEnabled' => $editorEnabled
+    ]);
+}
 ?>
 
 <!-- Edit button -->
@@ -25,4 +27,3 @@ if($hasSurveyContentUpdatePermission) {
     );
 }
 ?>
-
